make sure model evaluation is done in double precision

+++ src/pysm3/models/cmb.py
@@ -54,7 +54,7 @@ class CMBMap(Model):
         # do not use normalize weights because that includes a transformation
         # to spectral radiance and then back to RJ
         if weights is None:
-            weights = np.ones(len(freqs), dtype=np.float32)
+            weights = np.ones(len(freqs), dtype=np.float64)
 
         scaling_factor = utils.bandpass_unit_conversion(
             freqs * u.GHz, weights, output_unit=u.uK_RJ, input_unit=u.uK_CMB
